+++ app/src/main/java/com/android/example/motoapp/Table.kt
@@ -4,6 +4,7 @@ import android.location.Location
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.time.LocalTime
 
 @Entity(tableName = "location_table")
 class Table (
@@ -12,7 +13,7 @@ class Table (
     @ColumnInfo(name = "latitude") val latitudeDB:Double,
     @ColumnInfo(name = "longitude") val longitudeDB:Double,
     @ColumnInfo(name="rotation")val rotationDB:String,
-    @ColumnInfo(name = "time") val timeDB:Long
+    @ColumnInfo(name = "time") val timeDB:Int
 )
